Moved voting logic to its own file
